refactor: Cart 도메인 모델에서 페이징 로직 분리 및 CartViewModel로 이동
