+++ app/src/test/java/woowacourse/shopping/model/CartTest.kt
@@ -49,44 +49,6 @@ class CartTest {
         }.isInstanceOf(IllegalArgumentException::class.java)
     }
 
-    @Test
-    fun `페이지 사이즈에 맞춰 아이템을 반환한다`() {
-        val cart = createCart(size = 6)
-
-        val result = cart.getPage(page = 0, pageSize = 5)
-
-        assertThat(result.items.map { it.product.id })
-            .containsExactly("1", "2", "3", "4", "5")
-        assertThat(result.page).isEqualTo(0)
-        assertThat(result.isCanMoveNext).isTrue()
-    }
-
-    @Test
-    fun `남은 아이템이 한 페이지의 적정 아이템 개수보다 적을 경우, 남은 아이템만 반환한다`() {
-        val cart = createCart(size = 14)
-
-        val result = cart.getPage(page = 2, pageSize = 5)
-
-        assertThat(result.items.map { it.product.id })
-            .containsExactly("11", "12", "13", "14")
-        assertThat(result.page).isEqualTo(2)
-        assertThat(result.isCanMoveNext).isFalse()
-    }
-
-    @Test
-    fun `마지막 페이지에서는 다음 페이지 이동이 불가능하다`() {
-        val cart = createCart(size = 10)
-
-        val result = cart.getPage(page = 2, pageSize = 5)
-
-        assertThat(result.isCanMoveNext).isFalse()
-    }
-
-    private fun createCart(size: Int): Cart =
-        (1..size).fold(Cart()) { cart, id ->
-            cart.addItem(createProduct(id = id.toString()))
-        }
-
     private fun createProduct(id: String): Product =
         Product(
             id = id,
